Big refactor: package service depends on package utils
Previously, package utils depended on package service.  I didn't like how that made the code dependency tree work, so I refactored so that all packages depend on utils, since that makes the most sense to me.  Not sure exactly what to do with CommandEnvironment and EnvironmentMapper, but I've moved them to utils for now.