Huge rewrite of token-push notifications handling so service.Configs don't have notificationsChan, and thus package service does not depend on package notifications
Introduced a middleware layer (cmd/token-push/notificationsManager.go), that contains listeners, aggregators, and sorters to send notifications to the right place.  In this rewrite, when a serviceConfig is created, the code also registers a new notifications channel in memory.  A given worker will send all of its notifications on a single channel which the startServiceConfigWorkerForProcessing now also returns to main().  At the same time, startServiceConfigWorkerForProcessing will spin up a listener that reads from its returned notifications channel and sends the data into the aggregation channel, notificationsFromWorkersChan. A separate sorter, directNotificationsToManagers, listens on this latter chan, and sorts the notifications depending on the registration of the notifications channels.

This has two big effects:  1)  It gets rid of the unrelated field NotificationsChan from the service.Config object, and 2)  it thus removes the dependency of package service on package notifications, which is a much cleaner design.