Changed return values of IsSliceSubslice if the test failed
If IsSliceSubslice ascertains that sliceOne is not a subslice of sliceTwo, it should not
return an error as it previously did.  The test itself ran without issue, and an error
should only be returned if the test could not be run.  In the case where the test ran but
failed, the proper return tuple is (testResult, errorWhileRunningTest) = (false, nil).