Support HTGETTOKENOPTS env var and enforce minimum token lifetime
We now read HTGETTOKENOPTS from the environment and use that as long as it has the correct --credkey. If it does not, we add it onto the given HTGETTOKENOPTS. We also now enforce that any obtained vault token must have at least a certain lifetime (default 3 days).  This value is configurable (minTokenLifetime).