Fixed concurrency issue in 0.9 series with multiple concurrent vaultToken.StoreAndValidateToken calls
The utility that this func calls for both current TokenStorer implementations is condor_vault_storer.  condor_vault_storer is not concurrent safe for multiple calls for the same vault token from the same client.  Thus, we can concurrently call condor_vault_storer for different services, but not for the same service. Unfortunately, this leads to a significant slowdown, but we'll have to live with it to avoid race conditions that were present in 0.9 and 0.9.1, and only visible at scale.