Changes to types scheddCache and scheddCacheEntry
1. scheddCache: Converted the cache field to a
   map[string]*scheddCacheEntry and added a mutex to the type so that we
could lock down the globalScheddCache - which is a variable of type
scheddCache. I did this because we were running into race conditions
where sync.Map was allowing multiple readers which were reading and
writing to the same key (collector host), and sync.Map isn't equipped
for that.

2. scheddCacheEntry: Added a field err of type error that stores whether
   or not an error was encountered while populating the cache entry.
This allows us to create a scheddCacheEntry that is indicated as being
"bad", which allows us to skip over it if needed.