Huge rewrite.  Looping over schedds moved to StoreAndGetTokenWorker as well as new StoreAndGetTokensInteractiveWorker
In discussion with stlammel about this code base, he brought up that it
wasn't easy to find where exactly these workers loop over the schedds to
store tokens.  To make that much clearer, I moved the loops out of
StoreAndGetTokenForSchedds, renamed that function to
StoreAndGetTokenForSchedd, and made it only handle the operations for a
single token store.  The loop then moved into the two worker funcs
mentioned above.  This required some rewriting of the error handling as
well, since in both cases, we need to check for and report the correct
errors across different operations.